refactor: custom pathing for citizenship db funcs
